// VIS_Application/app/Model/Notification/TicketListClosed.ts
export interface ITicketListClosed {
    CompanyId: number;
    Mode: string;
    Id: number;
    Subject: string;
    CreatedOn: Date;
    AddressToGroup: string;
    AssignTo: string;
    CreatedByName: string;
    UpdatedByName: string;
    UpdatedOn: Date;
    Priority: string;
    Status: string;
    CreatedBy: number;
    UpdatedBy: number
    IsActive: boolean;
    UserId: number;
    FileName: string;

}